+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Enhancement:
 - Added `Publish-VersionHealth` to return the version health from the SoS Health Summary JSON data.
 - Updated `Invoke-VcfHealthReport` to include the version health using the `Publish-VersionHealth` cmdlet.
 - Added `Show-ReportingOutput` cmdlet to format output to the console when `PowerVCF` is not installed. [GH-121](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-reporting/pull/121)
+- Updated `Publish-VsanHealth` to include the results for capacity utilization and the active resysc of objects.
 
 Refactor: